Improve list and detail page markup
