+++ public/index.html
@@ -19,7 +19,7 @@
     </noscript>
 
     <div class="d-flex flex-column flex-md-row align-items-center p-3 px-md-4 mb-3 header border-bottom box-shadow">
-      <h5 class="my-0 mr-md-auto">Green Spider</h5>
+      <h5 class="my-0 mr-md-auto"><a href="/">Green Spider</a></h5>
       <nav class="my-2 my-md-0 mr-md-3">
         <span id="status" class="p-2">Lade Daten...</span>
         <a class="p-2 text-light" href="https://github.com/netzbegruenung/green-spider/"><i class="ion-logo-github"></i> GitHub</a>
