+++ src/ResultsList.css
@@ -1,17 +1,23 @@
 .TypeField {
-  font-size: 0.8em;
+  font-size: 85%;
   font-weight: 700;
   text-decoration-line: none;
   color: #333;
 }
 
 .ResultsList {
-  padding: 3px 0px;
-  margin: 6px 0px;
+  font-size: 1.3rem;
+  padding: 0.2rem 0;
+  margin: 0.3rem 0;
+  border-radius: 0.3rem;
 }
 
-.ResultList:hover {
-  background-color: #ccc;
+div.ResultsList:hover {
+  background-color: #eee;
+}
+
+a.ResultsList:hover {
+  text-decoration: none;
 }
 
 .ScoreField {
@@ -19,7 +25,7 @@
   background-color: #ccc;
   padding: 6px 8px;
   display: inline-block;
-  width: 3.5em;
+  width: 5em;
   text-align: center;
   font-family: 'Arvo Gruen', 'Arvo', sans-serif;
   font-size: 120%;
