+++ src/ResultsList.js
@@ -4,6 +4,7 @@
 
 import React, { Component } from 'react';
 import { Link } from "react-router-dom";
+import LocationLabel from './LocationLabel';
 import './ResultsList.css';
 import punycode from 'punycode';
 import chroma from 'chroma-js';
@@ -13,112 +14,7 @@ class ScoreField extends Component {
   render() {
     var ratio = this.props.score / this.props.maxScore;
     var bg = chroma.mix('#8D5335', '#75B66B', ratio);
-    return <span className='ScoreField' style={{backgroundColor: bg.hex()}}>{ this.props.score }</span>;
-  }
-}
-
-class StateField extends Component {
-  render() {
-    var label = this.props.state;
-
-    switch (this.props.state) {
-      case 'Nordrhein-Westfalen':
-        label = 'NW';
-        break;
-      case 'Rheinland-Pfalz':
-        label = 'RP';
-        break;
-      case 'Niedersachsen':
-        label = 'NS';
-        break;
-      case 'Baden-Württemberg':
-        label = 'BW';
-        break;
-      case 'Bayern':
-        label = 'BY';
-        break;
-      case 'Berlin':
-        label = 'BE';
-        break;
-      case 'Brandenburg':
-        label = 'BB';
-        break;
-      case 'Bremen':
-        label = 'HB';
-        break;
-      case 'Hamburg':
-        label = 'HB';
-        break;
-      case 'Hessen':
-        label = 'HE';
-        break;
-      case 'Mecklenburg-Vorpommern':
-        label = 'MV';
-        break;
-      case 'Saarland':
-        label = 'SL';
-        break;
-      case 'Sachsen':
-        label = 'SN';
-        break;
-      case 'Sachsen-Anhalt':
-        label = 'SA';
-        break;
-      case 'Schleswig-Holstein':
-        label = 'SH';
-        break;
-      case 'Thüringen':
-        label = 'SN';
-        break;
-      default:
-        label = this.props.state;
-    }
-    
-    return <abbr title={this.props.state}>{label}</abbr>;
-  }
-}
-
-class TypeField extends Component {
-  render() {
-    var label = '';
-    var title = '';
-
-    if (this.props.level === 'DE:BUNDESVERBAND') {
-      if (this.props.type === 'YOUTH_ORGANIZATION') {
-        title = 'Grüne Jugend Bundesverband';
-        label = 'GJ BV';
-      } else {
-        title = 'Bundesverband';
-        label = 'BV';
-      }
-    } else if (this.props.level === 'DE:LANDESVERBAND') {
-      if (this.props.type === 'YOUTH_ORGANIZATION') {
-        title = 'Grüne Jugend Landesverband';
-        label = 'GJ LV';
-      } else {
-        title = 'Landesverband';
-        label = 'LV';
-      }
-    } else if (this.props.level === 'DE:REGIONALVERBAND') {
-      title = 'Regionalverband';
-      label = 'RV';
-    } else if (this.props.level === 'DE:BEZIRKSVERBAND') {
-      title = 'Bezirksverband';
-      label = 'BeV';
-    } else if (this.props.level === 'DE:KREISVERBAND') {
-      if (this.props.type === 'YOUTH_ORGANIZATION') {
-        title = 'Grüne Jugend Kreisverband';
-        label = 'GJ KV';
-      } else {
-        title = 'Kreisverband';
-        label = 'KV';
-      }
-    } else if (this.props.level === 'DE:ORTSVERBAND') {
-      title = 'Ortsverband';
-      label = 'OV';
-    }
-
-    return <abbr className='TypeField' title={title}>{label}</abbr>;
+    return <span className='ScoreField align-self-center' style={{backgroundColor: bg.hex()}}>{ this.props.score }</span>;
   }
 }
 
@@ -130,39 +26,10 @@ class URLField extends Component {
     }
     return null;
   }
-  
-  // truncation for too long URLs
-  trunc(s, length) {
-    if (s.length > length) {
-      s = s.substring(0, length) + '…';
-    }
-    return s;
-  }
 
   render() {
-    var labelURL = this.trunc(this.displayURL(punycode.toUnicode(this.props.inputURL)), 40);
-    return <span className='URLField'>{ labelURL }</span>;
-  }
-}
-
-class LocationLabel extends Component {
-  render() {
-    var label = "";
-    var type = <TypeField level={this.props.level} type={this.props.type} />;
-    
-    if (this.props.city === null && this.props.district !== null) {
-      label = this.props.district;
-      return <span className='LocationLabel'>{type} {label}, <StateField state={this.props.state} /></span>;
-    } else if (this.props.city !== null && this.props.district === null) {
-      label = this.props.city;
-      return <span className='LocationLabel'>{type} {label}, <StateField state={this.props.state} /></span>;
-    } else if (this.props.city !== null && this.props.district !== null) {
-      label = `${this.props.city}, ${this.props.district}`;
-      return <span className='LocationLabel'>{type} {label}, <StateField state={this.props.state} /></span>;
-    }
-    
-    label = this.props.state;
-    return <span className='LocationLabel'>{type} {label}</span>;
+    var labelURL = this.displayURL(punycode.toUnicode(this.props.inputURL));
+    return <span className='URLField text-truncate'>{ labelURL }</span>;
   }
 }
 
@@ -178,13 +45,13 @@ class ResultsList extends Component {
     this.props.results.forEach((element, index) => {
 
       var row = (
-        <Link key={element.input_url} to={`/sites/${ encodeURIComponent(element.input_url) }`}>
+        <Link key={element.input_url} to={`/sites/${ encodeURIComponent(element.input_url) }`} className='ResultsList'>
           <div className='ResultsList row'>
-            <div className='col-9'>
-              <LocationLabel level={element.meta.level} type={element.meta.type} district={element.meta.district} city={element.meta.city} state={element.meta.state} />
+            <div className='col-9 col-sm-10 col-md-10'>
+              <LocationLabel level={element.meta.level} type={element.meta.type} district={element.meta.district} city={element.meta.city} state={element.meta.state} truncate={true} />
               <URLField inputURL={element.input_url} canonicalURLs={element.resulting_urls} />
             </div>
-            <div className='col-3'>
+            <div className='col-3 col-sm-2 col-md-2 d-flex'>
               <ScoreField score={element.score} maxScore={13} />
             </div>
           </div>
